refactor: 레이아웃 형식 변경
- 레이아웃 구조가 잘못되서 싹 다 수정
- 바꾸면서 이상해진부분이 있으면 말해주세요
