+++ client/src/App.tsx
@@ -1,15 +1,6 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'lib/router';
-import {
-  MainPage,
-  UserPage,
-  NotFoundPage,
-  SagaCounterPage,
-  TestPage,
-  SagaAxiosPage,
-  LoginPage,
-  SignupPage,
-} from 'pages';
+import { MainPage, NotFoundPage, SagaCounterPage, TestPage, LoginPage, SignupPage } from 'pages';
 import Theme from './styles/theme';
 
 const App: React.FC = () => {
@@ -18,12 +9,10 @@ const App: React.FC = () => {
       <BrowserRouter>
         <Switch>
           <Route exact path="/" component={MainPage} />
-          <Route exact path="/login" component={LoginPage} />
-          <Route exact path="/signup" component={SignupPage} />
-          <Route exact path="/user" component={UserPage} />
           <Route exact path="/test" component={TestPage} />
           <Route exact path="/counter" component={SagaCounterPage} />
-          <Route exact path="/axios" component={SagaAxiosPage} />
+          <Route exact path="/login" component={LoginPage} />
+          <Route exact path="/signup" component={SignupPage} />
           <Route path="" component={NotFoundPage} />
         </Switch>
       </BrowserRouter>
