+++ client/src/components/common/__test__/footer.test.tsx
