+++ client/src/components/common/__test__/header.test.tsx
