+++ client/src/components/common/footer.tsx
