+++ client/src/components/common/header.tsx
