+++ client/src/components/index.ts
@@ -1,5 +1,3 @@
 export { default as Logo } from './common/logo';
 export { default as Navbar } from './common/navbar';
 export { default as Modal } from './common/modal';
-
-export { default as Layout } from './layout';
