+++ client/src/pages/index.ts
@@ -1,8 +1,6 @@
 export { default as MainPage } from './main-page';
-export { default as UserPage } from './user-page';
 export { default as NotFoundPage } from './not-found-page';
 export { default as TestPage } from './test-page';
 export { default as SagaCounterPage } from './saga-counter-page';
-export { default as SagaAxiosPage } from './saga-axios-page';
 export { default as LoginPage } from './login-page';
 export { default as SignupPage } from './signup-page';
