--- client/src/pages/saga-axios-page.tsx
@@ -1,8 +1,0 @@
-import AuthContainer from 'containers/axios-container';
-import React, { FC } from 'react';
-
-const SagaAxiosPage: FC = () => {
-  return <AuthContainer />;
-};
-
-export default SagaAxiosPage;
