--- client/src/pages/user-page.tsx
@@ -1,8 +1,0 @@
-import React, { ReactElement } from 'react';
-import { Layout } from 'components';
-
-const UserPage = (): ReactElement => {
-  return <Layout>This is UserPage</Layout>;
-};
-
-export default UserPage;
